chore(middleware): rename middleware folder to use plural
